CW3M_McKenzie.envx - Add studyAreaName='McKenzie' to the <settings> block. Flow_McKenzie.xml - Reduce Clear Lake springs from 5.7 cms to 32 cfs. Reduce Smith R springs from 0.76 cms to 4 cfs. Reduce S Fork springs above CGR from 6.06 cms to 139 cfs. Take the difference between the simulated area drained and gage area drained into account in the Clear Lake column in the flow skill assessment reports. PEST_ClearLake46.xml - Decrease the Clear Lake springs from 6.6 cms to 32 cfs. APs.cpp - Fix a compiler warning. EnvContext.h - Add m_studyAreaName to class EnvContext.

+++ SourceCode/APs/APs.cpp
@@ -983,7 +983,7 @@ BOOL WW2100AP::InitRun( EnvContext *pContext, bool useInitSeed )
 
 bool WW2100AP::InitRunGetWeather(EnvContext* pContext)
 {
-   if (m_stationaryClimSeed != 0) m_WeatherRandomDraw.SetSeed(m_stationaryClimSeed);
+   if (m_stationaryClimSeed != 0) m_WeatherRandomDraw.SetSeed((long)m_stationaryClimSeed);
 
    return(true);
 } // end of WW2100AP::InitRunGetWeather()
